Avoid page margin hacks
This undoes the hack to force a zero page margin in order to hide
the default headers and footers. This tunred out to only work for
single page documents. Instead the extension now opens the print
preview and lets the user configure the headers and footers on their
own.

Fixes #1